complete timer1 interrupt
require test on real board cuz the QEMU may not support

+++ Task-2/src/gpio.h
@@ -5,7 +5,7 @@
 */
 #define RPI3 //enable when using RPI3 (QEMU emulation/ real board)
 #ifdef RPI3 //RPI3
-    #define MMIO_BASE 0x3F000000
+    #define MMIO_BASE (unsigned int)0x3F000000
 #else //RPI4
     #define MMIO_BASE 0xFE000000
 #endif
